Extracting component logic
